requirements.txt
setuptools>=69.0.2
Pillow>=10.1.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31
Flask-Login==0.6.3
Flask-Mail
Flask-Migrate
Flask-WTF
Flask-Caching
SQLAlchemy
qrcode
razorpay
python-dotenv
Werkzeug
itsdangerous
setuptools
psycopg2-binary
gunicorn
num2words
reportlab
pytz
APScheduler==3.10.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0